httpHeartbeat and retry-after support
I added a `heartbeatURL` property to config.json, which is an HTTP/HTTPS
url to which the bot will make a request every 5 minutes. This can be
used for uptime monitoring.

I changed `fetchStatusCode.mjs` to `fetchUtils.mjs`, which better
reflects the toolbox of HTTP functions it has become.

I also added support for retry-after in `fetchResponseChain`, so it can
automatically retry requests which respond with 429 Too Many Requests or
503 Service Unavailable.