.resume() and even more logging
I learned that when you make a request using http.get, in your response
handler, you're supposed to either listen to the data event of the
response object, or call `response.resume()` to discard the data, in
order for the request to actually end if you don't do that, the request
object gets held in memory indefinitely, for some reason.
https://nodejs.org/api/http.html#class-httpclientrequest

To address this, I added a `.resume()` call to `fetchInitialResponse()`.

I fixed the regexes used in `httpMonitor.mjs` for `splitMessageRegex` to
actually split messages where I intended.

I added a couple more console logs for even more debugging power (idk).