Complete manageTwitch.mjs command (first draft)
I implemented all the subcommands for /manage-twitch. For the most part,
this basically just means calling the relevant function in
manageTwitchUtils.mjs, which was my intention.

I also made a minor change in the /wiki command to parse out left angle
brackets `<`, which were causing errors in the MediaWiki API.