Fixing bug when single message is bulkDeleted
I previously tried to fix this bug, but forgot to return from the catch
statement. Also, I updated the readme a bit, particularly to remove the
lines for things I had already accomplished. Finally, I updated the
example value in the twitch guild-config-schema to actually be correct
for streamsChannelMessages, which should map to objects rather than
arrays.