Routine error reporting, restart command
I noticed that the monitorTwitchStreams routine errored out with a
FetchError. The stack trace doesn't include the source, which is weird,
but based on the API endpoint URL included in the message, I can tell
that it came from the `getStreamsManyIds` async generator. I added some
logic there to redo requests a few times when they fail, but I'm not
100% sure it's a fix. To compensate for this, and other past issues with
routines, I've added some logic to index.mjs to send a discord message
to the development guild whenever a routine fails, so I don't have to
SSH into the server to check for routine errors. Finally, I added a
restart command, which simply exits the process, after which PM2 should
restart it.