Adding /is-twitch-highlight command functionality
Implementing a command for getting info about twitch VODs, particularly
whether or not they are a highlight (i.e. permanently hosted) as opposed
to an archived video (AKA a Past Broadcast) which is automatically
deleted after a while. The main purpose of this is for quickly checking
whether a VOD in a leaderboard submission is valid.