Twitch monitoring - Change from tags to keywords
Previously, Autoride would filter streams of followed games based on the
tags API. On the 01st of August, 2022, Twitch updated the tag system to
allow free-form/custom tags
(https://blog.twitch.tv/en/2022/08/01/custom-tags-discovery-on-your-terms/).
Unfortunately, they did not update the API with this update. Custom tags
can't be inspected through the API at all. Pre-set tags still exist in
the API, but they cannot be set through the stream dashboard. The only
continuing use for these is for Language tags, which still get set and
returned in the API. I've filed a UserVoice idea to fix this
(https://twitch.uservoice.com/forums/310210-discover/suggestions/45656053-api-support-for-custom-tags),
but in the meantime, I've changed the bot to check the stream title for
a list of case-insensitive keywords. At least one of the keywords must
be present in the title, and not as a part of another word (e.g. the
keyword "AB" will not match "About time"), for the stream to be
reported. All tag-related functionality and subcommands have been
removed and replaced with analogous keyword ones. If and when custom tag
support is added to the API, I will consider switching back or using
both.