Automatically restart if error reporting fails
Previously, the bot would try to send a message if a command failed, and
send a message to the dev server if a routine failed. Unfortunately,
both of these kinds of errors often occur due to network problems, which
also cause the subsequent error message to fail to send to Discord. This
ends up making the error silent. To mitigate this, I've made it so the
bot restarts itself when error-reporting messages fail to send.