Converting setDefaultPermission to setDefaultMemberPermissions
setDefaultPermission was deprecated a while ago with the new integration
permission system in Discord. It no longer works. I'm replacing it with
`setDefaultMemberPermissions(0)`, which makes it so that by default you
must have the "administrator" role permission in order to use the given
commands. You can manually assign other roles to it, but this is a good
default because @borks in the Okami discord is an admin role. You can
also use this to make it so you need an existing role permission, like
"ban users" or "delete messages" to use a command, which might make
sense for certain commands like `prune`, but it makes no difference to
me since this bot is only in 2 servers right noww.