Initial On this day implementation and etc.
I think I've done all the baseline work for the on-this-day feature to
work, but I haven't done any debugging yet. I probably should've done
more commits up to this point, but oh well. I did a few small tweaks to
other files, such as adding types to all of the command execute
functions.

There are 4 on-this-day-related files:
- manageOTD.mjs - The command handler for adding, listing, removing,
and posting notable events.
- onThisDay.mjs - The AJV schema for the guild config JSON file used to
store events and announcement posting data. The events are organized by
month > day > year, which feels weird, but turns out to be appropriate
because we want to find all events on the same month and day primarily,
since we want to be able to easily find all events on that day across
every year.
- postOTD.mjs - A routine which is only meant to run once per day,
just past midnight in local time. The naive approach would have been to
set the interval to 24 hours, but this would break if the bot ever
reset, either causing the next announcement to be missed or causing the
previous announcement to be repeated. Instead, I store the last date an
announcement was posted in the guild config, and check every 10 minutes
if the current date is different yet.
- otdUtils.mjs - This contains all of the functions involved with
manipulating the guild config and posting announcements.