OTD Functionality fixes
I did a bunch of bugfixing to get the on-this-day functionality working
more-or-less as desired. One major change is that OTD announcements are now
posted via embeds. This ensures that markdown features like masked links always
work as intended. This also allows us to add a fun accent color and thumbnail.
For the color, I made it rotate through the rainbow over the course of the
year, and cycle up and down in saturation over the course of the 12 year zodiac
cycle. For the thumbnail, I used the steam achievement icons for Okami. Of
course, this makes this no-longer generic for different content, but it would
not be too hard to change out those features in the code if it was ever
desired.