postGuildOTDAnnouncement returns earlier
Previously, if an announcement hadn't been posted today,
`postGuildOTDAnnouncement` would get the Discord channel API object
first before checking whether there were actually any announcements to
post. This change makes it so that it checks if there are any
announcements first, and only gets the channel API object if there are
some. This reduces the chance of unnecessary API calls to get the channel
object.